+++ package.json
@@ -1,73 +1,73 @@
 {
-    "name": "@elysiajs/opentelemetry",
-    "version": "1.3.0",
-    "license": "MIT",
-    "scripts": {
-        "dev": "bun run --watch example/index.ts",
-        "test": "bun test && npm run test:node",
-        "test:node": "npm install --prefix ./test/node/cjs/ && npm install --prefix ./test/node/esm/ && node ./test/node/cjs/index.js && node ./test/node/esm/index.js",
-        "build": "bun build.ts",
-        "release": "npm run build && npm run test && npm publish --access public"
-    },
-    "author": {
-        "name": "saltyAom",
-        "url": "https://github.com/SaltyAom",
-        "email": "saltyaom@gmail.com"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/elysiajs/opentelemetry"
-    },
-    "main": "./dist/cjs/index.js",
-    "module": "./dist/index.mjs",
-    "types": "./dist/index.d.ts",
-    "exports": {
-        "./package.json": "./package.json",
-        ".": {
-            "types": "./dist/index.d.ts",
-            "import": "./dist/index.mjs",
-            "require": "./dist/cjs/index.js"
-        }
-    },
-    "keywords": [
-        "elysia",
-        "opentelemetry",
-        "tracing"
-    ],
-    "dependencies": {
-        "@opentelemetry/api": "^1.9.0",
-        "@opentelemetry/instrumentation": "^0.200.0",
-        "@opentelemetry/sdk-node": "^0.200.0"
-    },
-    "devDependencies": {
-        "@axiomhq/js": "^1.3.1",
-        "@elysiajs/eden": "^1.2.0",
-        "@elysiajs/graphql-yoga": "^1.2.0",
-        "@envelop/core": "^5.2.3",
-        "@envelop/opentelemetry": "^7.1.3",
-        "@opentelemetry/context-async-hooks": "^2.0.0",
-        "@opentelemetry/exporter-jaeger": "^2.0.0",
-        "@opentelemetry/exporter-metrics-otlp-proto": "^0.200.0",
-        "@opentelemetry/exporter-trace-otlp-http": "^0.200.0",
-        "@opentelemetry/exporter-trace-otlp-proto": "^0.200.0",
-        "@opentelemetry/sdk-metrics": "^2.0.0",
-        "@opentelemetry/sdk-trace-node": "^2.0.0",
-        "@opentelemetry/sdk-trace-web": "^2.0.0",
-        "@types/bun": "1.2.10",
-        "elysia": "1.3.0-exp.67",
-        "eslint": "9.25.1",
-        "tsup": "^8.4.0",
-        "typescript": "^5.8.3"
-    },
-    "peerDependencies": {
-        "elysia": ">= 1.3.0"
-    },
-    "peerDependenciesMeta": {
-        "graphql": {
-            "optional": true
-        },
-        "graphql-yoga": {
-            "optional": true
-        }
-    }
+	"name": "@elysiajs/opentelemetry",
+	"version": "1.3.0",
+	"license": "MIT",
+	"scripts": {
+		"dev": "bun run --watch example/index.ts",
+		"test": "bun test --preload ./test/test-setup.ts && npm run test:node",
+		"test:node": "npm install --prefix ./test/node/cjs/ && npm install --prefix ./test/node/esm/ && node ./test/node/cjs/index.js && node ./test/node/esm/index.js",
+		"build": "bun build.ts",
+		"release": "npm run build && npm run test && npm publish --access public"
+	},
+	"author": {
+		"name": "saltyAom",
+		"url": "https://github.com/SaltyAom",
+		"email": "saltyaom@gmail.com"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/elysiajs/opentelemetry"
+	},
+	"main": "./dist/cjs/index.js",
+	"module": "./dist/index.mjs",
+	"types": "./dist/index.d.ts",
+	"exports": {
+		"./package.json": "./package.json",
+		".": {
+			"types": "./dist/index.d.ts",
+			"import": "./dist/index.mjs",
+			"require": "./dist/cjs/index.js"
+		}
+	},
+	"keywords": [
+		"elysia",
+		"opentelemetry",
+		"tracing"
+	],
+	"dependencies": {
+		"@opentelemetry/api": "^1.9.0",
+		"@opentelemetry/instrumentation": "^0.200.0",
+		"@opentelemetry/sdk-node": "^0.200.0"
+	},
+	"devDependencies": {
+		"@axiomhq/js": "^1.3.1",
+		"@elysiajs/eden": "^1.2.0",
+		"@elysiajs/graphql-yoga": "^1.2.0",
+		"@envelop/core": "^5.2.3",
+		"@envelop/opentelemetry": "^7.1.3",
+		"@opentelemetry/context-async-hooks": "^2.0.0",
+		"@opentelemetry/exporter-jaeger": "^2.0.0",
+		"@opentelemetry/exporter-metrics-otlp-proto": "^0.200.0",
+		"@opentelemetry/exporter-trace-otlp-http": "^0.200.0",
+		"@opentelemetry/exporter-trace-otlp-proto": "^0.200.0",
+		"@opentelemetry/sdk-metrics": "^2.0.0",
+		"@opentelemetry/sdk-trace-node": "^2.0.0",
+		"@opentelemetry/sdk-trace-web": "^2.0.0",
+		"@types/bun": "1.2.10",
+		"elysia": "1.3.0-exp.67",
+		"eslint": "9.25.1",
+		"tsup": "^8.4.0",
+		"typescript": "^5.8.3"
+	},
+	"peerDependencies": {
+		"elysia": ">= 1.3.0"
+	},
+	"peerDependenciesMeta": {
+		"graphql": {
+			"optional": true
+		},
+		"graphql-yoga": {
+			"optional": true
+		}
+	}
 }
